Add donor parent/child displays
For clealiness, moved donor components from biosamples.js to a new donor.js files. Also added new parent/child human_donor local test data, and allowed for self-referencing human_donor objects so that parents, children, siblings, fraternal_twi, and idential_twin can index. Redmine #4885
