+++ src/encoded/static/components/biosample.js
@@ -507,424 +507,6 @@ var BiosampleTermId = React.createClass({
 });
 
 
-var HumanDonor = module.exports.HumanDonor = React.createClass({
-    propTypes: {
-        context: React.PropTypes.object.isRequired, // Donor being displayed
-        biosample: React.PropTypes.object, // Biosample this donor is associated with
-    },
-
-    render: function () {
-        const { context, biosample } = this.props;
-        var references = pubReferenceList(context.references);
-
-        // Render tags badges
-        var tagBadges;
-        if (context.internal_tags && context.internal_tags.length) {
-            tagBadges = context.internal_tags.map(tag => <img src={'/static/img/tag-' + tag + '.png'} alt={tag + ' tag'} />);
-        }
-
-        return (
-            <div>
-                <Panel>
-                    <PanelHeading>
-                        <h4>Donor information</h4>
-                    </PanelHeading>
-                    <PanelBody>
-                        <dl className="key-value">
-                            <div data-test="accession">
-                                <dt>Accession</dt>
-                                <dd>{biosample ? <a href={context['@id']}>{context.accession}</a> : context.accession}</dd>
-                            </div>
-
-                            {context.aliases.length ?
-                                <div data-test="aliases">
-                                    <dt>Aliases</dt>
-                                    <dd>{context.aliases.join(", ")}</dd>
-                                </div>
-                            : null}
-
-                            {context.organism.scientific_name ?
-                                <div data-test="species">
-                                    <dt>Species</dt>
-                                    <dd className="sentence-case"><em>{context.organism.scientific_name}</em></dd>
-                                </div>
-                            : null}
-
-                            {context.life_stage ?
-                                <div data-test="life-stage">
-                                    <dt>Life stage</dt>
-                                    <dd className="sentence-case">{context.life_stage}</dd>
-                                </div>
-                            : null}
-
-                            {context.age ?
-                                <div data-test="age">
-                                    <dt>Age</dt>
-                                    <dd className="sentence-case">{context.age}{context.age_units ? ' ' + context.age_units : null}</dd>
-                                </div>
-                            : null}
-
-                            {context.sex ?
-                                <div data-test="sex">
-                                    <dt>Sex</dt>
-                                    <dd className="sentence-case">{context.sex}</dd>
-                                </div>
-                            : null}
-
-                            {context.health_status ?
-                                <div data-test="health-status">
-                                    <dt>Health status</dt>
-                                    <dd className="sentence-case">{context.health_status}</dd>
-                                </div>
-                            : null}
-
-                            {context.ethnicity ?
-                                <div data-test="ethnicity">
-                                    <dt>Ethnicity</dt>
-                                    <dd className="sentence-case">{context.ethnicity}</dd>
-                                </div>
-                            : null}
-
-                            {context.dbxrefs && context.dbxrefs.length ?
-                                <div data-test="external-resources">
-                                    <dt>External resources</dt>
-                                    <dd><DbxrefList values={context.dbxrefs} /></dd>
-                                </div>
-                            : null}
-
-                            {references ?
-                                <div data-test="references">
-                                    <dt>References</dt>
-                                    <dd>{references}</dd>
-                                </div>
-                            : null}
-
-                            {tagBadges ?
-                                <div className="tag-badges" data-test="tags">
-                                    <dt>Tags</dt>
-                                    <dd>{tagBadges}</dd>
-                                </div>
-                            : null}
-                        </dl>
-                    </PanelBody>
-                </Panel>
-            </div>
-        );
-    }
-});
-
-globals.panel_views.register(HumanDonor, 'HumanDonor');
-
-
-var MouseDonor = module.exports.MouseDonor = React.createClass({
-    render: function() {
-        var context = this.props.context;
-        var biosample = this.props.biosample;
-        var donorUrlDomain;
-        var references = pubReferenceList(context.references);
-
-        // Get the domain name of the donor URL
-        if (biosample && biosample.donor && biosample.donor.url) {
-            var donorUrl = url.parse(biosample.donor.url);
-            donorUrlDomain = donorUrl.hostname || '';
-        }
-
-        // Render tags badges
-        var tagBadges;
-        if (context.internal_tags && context.internal_tags.length) {
-            tagBadges = context.internal_tags.map(tag => <img src={'/static/img/tag-' + tag + '.png'} alt={tag + ' tag'} />);
-        }
-
-        return (
-            <div>
-                <Panel>
-                    <PanelHeading>
-                        <h4>Strain information</h4>
-                    </PanelHeading>
-                    <PanelBody>
-                        <dl className="key-value">
-                            <div data-test="accession">
-                                <dt>Accession</dt>
-                                <dd>{biosample ? <a href={context['@id']}>{context.accession}</a> : context.accession}</dd>
-                            </div>
-
-                            {context.aliases.length ?
-                                <div data-test="aliases">
-                                    <dt>Aliases</dt>
-                                    <dd>{context.aliases.join(", ")}</dd>
-                                </div>
-                            : null}
-
-                            {context.organism.scientific_name ?
-                                <div data-test="organism">
-                                    <dt>Species</dt>
-                                    <dd className="sentence-case"><em>{context.organism.scientific_name}</em></dd>
-                                </div>
-                            : null}
-
-                            {context.genotype ?
-                                <div data-test="genotype">
-                                    <dt>Genotype</dt>
-                                    <dd>{context.genotype}</dd>
-                                </div>
-                            : null}
-
-                            {context.mutated_gene && biosample && biosample.donor && biosample.donor.mutated_gene && biosample.donor.mutated_gene.label ?
-                                <div data-test="mutatedgene">
-                                    <dt>Mutated gene</dt>
-                                    <dd><a href={context.mutated_gene}>{biosample.donor.mutated_gene.label}</a></dd>
-                                </div>
-                            : null}
-
-                            {biosample && biosample.sex ?
-                                <div data-test="sex">
-                                    <dt>Sex</dt>
-                                    <dd className="sentence-case">{biosample.sex}</dd>
-                                </div>
-                            : null}
-
-                            {biosample && biosample.health_status ?
-                                <div data-test="health-status">
-                                    <dt>Health status</dt>
-                                    <dd className="sentence-case">{biosample.health_status}</dd>
-                                </div>
-                            : null}
-
-                            {donorUrlDomain ?
-                                <div data-test="mutatedgene">
-                                    <dt>Strain reference</dt>
-                                    <dd><a href={biosample.donor.url}>{donorUrlDomain}</a></dd>
-                                </div>
-                            : null}
-
-                            {context.strain_background ?
-                                <div data-test="strain-background">
-                                    <dt>Strain background</dt>
-                                    <dd className="sentence-case">{context.strain_background}</dd>
-                                </div>
-                            : null}
-
-                            {context.strain_name ?
-                                <div data-test="strain-name">
-                                    <dt>Strain name</dt>
-                                    <dd>{context.strain_name}</dd>
-                                </div>
-                            : null}
-
-                            {context.dbxrefs && context.dbxrefs.length ?
-                                <div data-test="external-resources">
-                                    <dt>External resources</dt>
-                                    <dd><DbxrefList values={context.dbxrefs} /></dd>
-                                </div>
-                            : null}
-
-                            {context.references && context.references.length ?
-                                <div data-test="references">
-                                    <dt>References</dt>
-                                    <dd>{pubReferenceList(context.references)}</dd>
-                                </div>
-                            : null}
-
-                            {tagBadges ?
-                                <div className="tag-badges" data-test="tags">
-                                    <dt>Tags</dt>
-                                    <dd>{tagBadges}</dd>
-                                </div>
-                            : null}
-                        </dl>
-
-                        {biosample && biosample.donor.characterizations && biosample.donor.characterizations.length ?
-                            <div>
-                                <hr />
-                                <h4>Characterizations</h4>
-                                <PanelBody addClasses="panel-body-doc-interior">
-                                    <DocumentsSubpanels documentSpec={{documents: biosample.donor.characterizations}} />
-                                </PanelBody>
-                            </div>
-                        : null}
-                    </PanelBody>
-                </Panel>
-            </div>
-        );
-    }
-});
-
-globals.panel_views.register(MouseDonor, 'MouseDonor');
-
-
-var FlyWormDonor = module.exports.FlyDonor = React.createClass({
-    render: function() {
-        var context = this.props.context;
-        var biosample = this.props.biosample;
-        var donorUrlDomain;
-
-        // Render tags badges
-        var tagBadges;
-        if (context.internal_tags && context.internal_tags.length) {
-            tagBadges = context.internal_tags.map(tag => <img src={'/static/img/tag-' + tag + '.png'} alt={tag + ' tag'} />);
-        }
-
-        return (
-            <div>
-                <Panel>
-                    <PanelHeading>
-                        <h4>Strain information</h4>
-                    </PanelHeading>
-                    <PanelBody>
-                        <dl className="key-value">
-                            <div data-test="accession">
-                                <dt>Accession</dt>
-                                <dd>{biosample ? <a href={context['@id']}>{context.accession}</a> : context.accession}</dd>
-                            </div>
-
-                            {context.aliases.length ?
-                                <div data-test="aliases">
-                                    <dt>Aliases</dt>
-                                    <dd>{context.aliases.join(", ")}</dd>
-                                </div>
-                            : null}
-
-                            {context.organism.scientific_name ?
-                                <div data-test="species">
-                                    <dt>Species</dt>
-                                    <dd className="sentence-case"><em>{context.organism.scientific_name}</em></dd>
-                                </div>
-                            : null}
-
-                            {context.genotype ?
-                                <div data-test="genotype">
-                                    <dt>Genotype</dt>
-                                    <dd>{context.genotype}</dd>
-                                </div>
-                            : null}
-
-                            {context.mutated_gene && biosample && biosample.donor && biosample.donor.mutated_gene && biosample.donor.mutated_gene.label ?
-                                <div data-test="mutatedgene">
-                                    <dt>Mutated gene</dt>
-                                    <dd><a href={context.mutated_gene['@id']}>{biosample.donor.mutated_gene.label}</a></dd>
-                                </div>
-                            : null}
-
-                            {biosample && biosample.sex ?
-                                <div data-test="sex">
-                                    <dt>Sex</dt>
-                                    <dd className="sentence-case">{biosample.sex}</dd>
-                                </div>
-                            : null}
-
-                            {biosample && biosample.health_status ?
-                                <div data-test="health-status">
-                                    <dt>Health status</dt>
-                                    <dd className="sentence-case">{biosample.health_status}</dd>
-                                </div>
-                            : null}
-
-                            {donorUrlDomain ?
-                                <div data-test="mutatedgene">
-                                    <dt>Strain reference</dt>
-                                    <dd><a href={biosample.donor.url}>{donorUrlDomain}</a></dd>
-                                </div>
-                            : null}
-
-                            {context.strain_background ?
-                                <div data-test="strain-background">
-                                    <dt>Strain background</dt>
-                                    <dd className="sentence-case">{context.strain_background}</dd>
-                                </div>
-                            : null}
-
-                            {context.strain_name ?
-                                <div data-test="strain-name">
-                                    <dt>Strain name</dt>
-                                    <dd>{context.strain_name}</dd>
-                                </div>
-                            : null}
-
-                            {context.dbxrefs && context.dbxrefs.length ?
-                                <div data-test="external-resources">
-                                    <dt>External resources</dt>
-                                    <dd><DbxrefList values={context.dbxrefs} /></dd>
-                                </div>
-                            : null}
-
-                            {tagBadges ?
-                                <div className="tag-badges" data-test="tags">
-                                    <dt>Tags</dt>
-                                    <dd>{tagBadges}</dd>
-                                </div>
-                            : null}
-                        </dl>
-                    </PanelBody>
-                </Panel>
-            </div>
-        );
-    }
-});
-
-globals.panel_views.register(FlyWormDonor, 'FlyDonor');
-globals.panel_views.register(FlyWormDonor, 'WormDonor');
-
-
-var Donor = module.exports.Donor = React.createClass({
-    render: function() {
-        var context = this.props.context;
-        var itemClass = globals.itemClass(context, 'view-item');
-        var altacc = context.alternate_accessions ? context.alternate_accessions.join(', ') : undefined;
-        var characterizationDocuments = [];
-        var donorDocuments = [];
-        var combinedDocuments = [];
-
-        // Collect the characterization documents
-        if (context.characterizations && context.characterizations.length) {
-            characterizationDocuments = context.characterizations;
-        }
-
-        // Collect the donor documents
-        if (context.documents && context.documents.length) {
-            donorDocuments = context.documents;
-        }
-
-        // Combine characterization and donor documents
-        combinedDocuments = [].concat(characterizationDocuments, donorDocuments);
-
-        // Set up breadcrumbs
-        var crumbs = [
-            {id: 'Donors'},
-            {id: <i>{context.organism.scientific_name}</i>}
-        ];
-
-        return (
-            <div className={itemClass}>
-                <header className="row">
-                    <div className="col-sm-12">
-                        <Breadcrumbs crumbs={crumbs} />
-                        <h2>{context.accession}</h2>
-                        {altacc ? <h4 className="repl-acc">Replaces {altacc}</h4> : null}
-                        <div className="status-line">
-                            <div className="characterization-status-labels">
-                                <StatusLabel title="Status" status={context.status} />
-                            </div>
-                        </div>
-                    </div>
-                </header>
-
-                <div>{PanelLookup(context)}</div>
-
-                <RelatedItems title={"Biosamples from this " + (context.organism.name == 'human' ? 'donor': 'strain')}
-                              url={'/search/?type=biosample&donor.uuid=' + context.uuid}
-                              Component={BiosampleTable} />
-
-                {combinedDocuments.length ?
-                    <DocumentsPanel documentSpecs={[{documents: combinedDocuments}]} />
-                : null}
-            </div>
-        );
-    }
-});
-
-globals.content_views.register(Donor, 'Donor');
-
-
 var Treatment = module.exports.Treatment = React.createClass({
     render: function() {
         var context = this.props.context;
